requirements.txt
requests
pandas
python-dotenv
scikit-learn
matplotlib
pyod